+++ libs/sdk-core/CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Improvements
 
 - Add `predictAnswer` method to `KnowledgeBox` class
+- Add support for `semantic-model-migrator` tasks
+- Support adding or remove semantic models
 
 # 1.24.1 (2025-02-05)